Generator: do not lock huginn_agent version in gemspec
The huginn_agent version in agent gems should not be locked in the gemspec per default, Huginn already has the version
locked in its Gemfile. As long as the Agent gem does not require a specific huginn_agent gem version it should not care
about it.
For the development of the Agent gem we specify the huginn_agent gem version in the Gemfile.