Make compatible with Rails 6 & 7 Zeitwerk loader
Rails 6 & 7 autoloading with Zeitwerk enabled didn't get along with this
gem's approach to loading agent gems and adding them to `Agent::TYPES`
when the `Agent` model got reloaded.

Add a `Zeitwerk::Loader` for each agent gem that is registered with
`Zeitwerk` so that Rails can load `Agents::YOUR_AGENT_CLASS` on demand.

For Rails "classic" autoloader, use `require_dependency` instead of
`require` so that the autoloader knows where to reload the agent class
from, as recommended in
https://guides.rubyonrails.org/v5.2/autoloading_and_reloading_constants.html#autoloading-and-require

Signed-off-by: Ross Williams <ross@ross-williams.net>